improving code and structure
