+++ src/models/todo/schema/todo.schema.ts
@@ -1,0 +1,18 @@
+import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
+import { HydratedDocument } from 'mongoose';
+
+export type TodoDocument = HydratedDocument<Todo>;
+
+@Schema({ timestamps: true })
+export class Todo {
+  @Prop({ required: true })
+  title: string;
+
+  @Prop()
+  description?: string;
+
+  @Prop({ default: false })
+  completed: boolean;
+}
+
+export const TodoSchema = SchemaFactory.createForClass(Todo);
