+++ src/models/todo/todo.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { TodoController } from './todo.controller';
-import { TodoService } from './todo.service';
+import { TodoController } from './controllers/todo.controller';
+import { TodoService } from './service/todo.service';
 
 @Module({
   controllers: [TodoController],
