--- src/todo/todo.schema.ts
@@ -1,18 +1,0 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { Document } from 'mongoose';
-
-export type TodoDocument = Todo & Document;
-
-@Schema()
-export class Todo {
-  @Prop({ required: true })
-  title: string;
-
-  @Prop()
-  description: string;
-
-  @Prop({ default: false })
-  isDone: boolean;
-}
-
-export const TodoSchema = SchemaFactory.createForClass(Todo);
